refactor: 데이터 수신 시 url을 같이 넘겨주도록 개선
