+++ src/api/api.ts
@@ -1,3 +1,5 @@
+import { CartItem, ProductItem, ReceivedCartItem } from "../types/types";
+
 export const fetchAddCart = async (url: string, id: number) => {
   const response = await fetch(`${url}/cart-items`, {
     method: "POST",
@@ -25,4 +27,32 @@ export const fetchUpdateCart = async (url: string, id: number, quantity: number)
   });
   const result = await response.json();
   console.log(result);
+};
+
+export const fetchCartList = async (url: string) => {
+  try {
+    const response = await fetch(`${url}/cart-items`);
+    const data = await response.json();
+    const checkedCartItems: CartItem[] = data.map(
+      (cartItem: ReceivedCartItem) => ({
+        ...cartItem,
+        checked: true,
+      })
+    );
+    return checkedCartItems;
+  } catch (error) {
+    console.error(error);
+    throw new Error();
+  }
+};
+
+export const fetchProductList = async (url: string) => {
+  try {
+    const response = await fetch("/products");
+    const data: ProductItem[] = await response.json();
+    return data;
+  } catch (error) {
+    console.error(error);
+    throw new Error();
+  }
 };
